Reflect BC setup and a concern
Reflect BC for velocities were set up in this install.

For Reflect BC, there is a vector field, grad_q, which
is for lagged pressure gradient. It is used in function
computeGradientQ_MAC_vd(), treated same as velocity field
a review will be needed for the future commit

removal tag: HAOZ should be removed in a phase where reflection bc
is tested.