RSRV steady state, special input and indices set up
This install is to ensure RSRV, which is also the code itself, is
capable of maintaining a steady state over time.

For steady state, PERIODIC requires much longer in Z direction to
achieve the goal.

setIndexMap() is to set up indices for the computational domain and
its buffer. This fix ***UGLY*** fixed the compatibility between
PERIODIC and REFLECTION BOUNDARY CONDITION.