Upgrade PETSc Wrapper for iFluid v3.5.4
This upgrade PETSc wrapper from 3.1(-p7,,p8) to PETSc 3.5.4.
New PETSc support quad precision which is one of the major reasons to
upgrade.
From now on, PETSc compile requires at least 3.2. For the sake of
simplicity, PETSc 3.5.4 is recommended. Alongside with a newer version
of GCC compiler (5.2.0) at least 4.60 for quad precision support.