Reduce Num_Iter Threshold for PETSc Solve
target function: compDiffWithSmoothProperty_velocity_MAC_coupled_vd()

solver.SetMaxIter(40000), the num_iter was decreased by a lot

The context is that continuing the update after introducing HYPRE
preconditioner "boomeramg" and this solver was not optimized by HYPRE by
no means.

Simply just reduce the number of iterations for the solver.